Report interfaces to a separate project
