+++ Services/WebStore.Interfaces/Service/IEmployeesData.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using WebStore.Domain.Entities;
 
-namespace WebStore.Infrastructure.Interfaces
+namespace WebStore.Interfaces.Service
 {
     public interface IEmployeesData
     {
