+++ Services/WebStore.Interfaces/Service/Interface.cs
@@ -1,6 +1,6 @@
 ﻿using WebStore.Domain.ViewModels;
 
-namespace WebStore.Infrastructure.Interfaces
+namespace WebStore.Interfaces.Service
 {
     public interface ICartService
     {
