+++ UI/WebStore/Components/SectionsViewComponent.cs
@@ -2,7 +2,7 @@
 using System.Linq;
 using Microsoft.AspNetCore.Mvc;
 using WebStore.Domain.ViewModels;
-using WebStore.Infrastructure.Interfaces;
+using WebStore.Interfaces.Service;
 
 namespace WebStore.Components
 {
