+++ UI/WebStore/Controllers/EmployeesController.cs
@@ -5,8 +5,8 @@ using Microsoft.AspNetCore.Mvc;
 using WebStore.Domain.Entities;
 using WebStore.Domain.Entities.Identity;
 using WebStore.Domain.ViewModels;
-using WebStore.Infrastructure.Interfaces;
 using WebStore.Infrastructure.Mapping;
+using WebStore.Interfaces.Service;
 
 namespace WebStore.Controllers
 {
