+++ UI/WebStore/Infrastructure/Services/InMemory/InMemoryEmployeesData.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using WebStore.Data;
 using WebStore.Domain.Entities;
-using WebStore.Infrastructure.Interfaces;
+using WebStore.Interfaces.Service;
 
 namespace WebStore.Infrastructure.Services.InMemory
 {
