+++ UI/WebStore/Infrastructure/Services/InSQL/SqlEmployeesData.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using WebStore.DAL.Context;
 using WebStore.Domain.Entities;
-using WebStore.Infrastructure.Interfaces;
+using WebStore.Interfaces.Service;
 
 namespace WebStore.Infrastructure.Services.InSQL
 {
