+++ UI/WebStore/Infrastructure/Services/InSQL/SqlProductData.cs
@@ -4,7 +4,7 @@ using Microsoft.EntityFrameworkCore;
 using WebStore.DAL.Context;
 using WebStore.Domain;
 using WebStore.Domain.Entities;
-using WebStore.Infrastructure.Interfaces;
+using WebStore.Interfaces.Service;
 
 namespace WebStore.Infrastructure.Services.InSQL
 {
