+++ UI/WebStore/Views/Home/Index.cshtml
@@ -1,5 +1,5 @@
-﻿@using WebStore.Infrastructure.Interfaces
-@using WebStore.Infrastructure.Mapping
+﻿@using WebStore.Infrastructure.Mapping
+@using WebStore.Interfaces.Service
 @inject IProductData ProductData
 @{
     var products = ProductData
