+++ WebStore.sln
@@ -15,6 +15,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebStore.DAL", "Services\We
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebStore", "UI\WebStore\WebStore.csproj", "{A9B58C64-D4CD-4C0B-9402-C76C4128C3D8}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebStore.Interfaces", "Services\WebStore.Interfaces\WebStore.Interfaces.csproj", "{875596F5-0DFD-48D6-B371-C609FE7E23C6}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -33,6 +35,10 @@ Global
 		{A9B58C64-D4CD-4C0B-9402-C76C4128C3D8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A9B58C64-D4CD-4C0B-9402-C76C4128C3D8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A9B58C64-D4CD-4C0B-9402-C76C4128C3D8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{875596F5-0DFD-48D6-B371-C609FE7E23C6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{875596F5-0DFD-48D6-B371-C609FE7E23C6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{875596F5-0DFD-48D6-B371-C609FE7E23C6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{875596F5-0DFD-48D6-B371-C609FE7E23C6}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -41,6 +47,7 @@ Global
 		{5638C20A-D706-4739-8C09-C6836BB6FCBD} = {90B918E8-0BDF-439C-AC9E-C163E2BDA03D}
 		{A1F2C90E-BE3D-4192-9DCE-E0351E4F0D42} = {F3B6B87B-E8CF-44E4-BB61-192F3F289A00}
 		{A9B58C64-D4CD-4C0B-9402-C76C4128C3D8} = {2BE43501-725B-49DC-91F9-154718C3B5C1}
+		{875596F5-0DFD-48D6-B371-C609FE7E23C6} = {F3B6B87B-E8CF-44E4-BB61-192F3F289A00}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {D180A988-D89A-45B3-81F7-DF7163BABFE3}
